test06_16.java
